Add usePreferences hook for handling preferences
Also remove the unused PreferencesProvider.jsx file.
